feat: allow clips submitted by many to rise up in queue
Now, clips will be queued based on the amount of people
who have tried to submit the clip (then the time they
were submitted). If a user is banned or timed out, then
they will only be removed as a submitter and the clip will
only be removed if others have not submitted.