feat(web/services): use twitch login name when possible
When the Twitch user details can be fetched use the login
name to ensure that we are connecting to chat properly.
When we cannot fetch the Twitch user details fallback to the
preferred username as it works in many cases. It specifically
does not seem to work when the user has a localized display
name.

closes: #418, #406

Signed-off-by: Jordan Shatford <jordanshatford@live.com>